Move pkg/fargate to pkg/actions
