+++ pkg/actions/fargate/fargate_test.go
@@ -8,20 +8,22 @@ import (
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 	"github.com/pkg/errors"
+	"github.com/weaveworks/eksctl/pkg/actions/fargate"
 	api "github.com/weaveworks/eksctl/pkg/apis/eksctl.io/v1alpha5"
 	"github.com/weaveworks/eksctl/pkg/eks/mocks"
-	"github.com/weaveworks/eksctl/pkg/fargate"
 	"github.com/weaveworks/eksctl/pkg/utils/retry"
 	"github.com/weaveworks/eksctl/pkg/utils/strings"
 )
 
 const clusterName = "non-existing-test-cluster"
 
+var retryPolicy = retry.NewTimingOutExponentialBackoff(5 * time.Minute)
+
 var _ = Describe("fargate", func() {
 	Describe("Client", func() {
 		Describe("CreateProfile", func() {
 			It("fails fast if the provided profile is nil", func() {
-				client := fargate.NewClient(clusterName, &mocks.EKSAPI{})
+				client := fargate.NewWithRetryPolicy(clusterName, &mocks.EKSAPI{}, &retryPolicy)
 				waitForCreation := false
 				err := client.CreateProfile(nil, waitForCreation)
 				Expect(err).To(HaveOccurred())
@@ -29,21 +31,21 @@ var _ = Describe("fargate", func() {
 			})
 
 			It("creates the provided profile without tag", func() {
-				client := fargate.NewClient(clusterName, mockForCreateFargateProfileWithoutTag())
+				client := fargate.NewWithRetryPolicy(clusterName, mockForCreateFargateProfileWithoutTag(), &retryPolicy)
 				waitForCreation := false
 				err := client.CreateProfile(createProfileWithoutTag(), waitForCreation)
 				Expect(err).To(Not(HaveOccurred()))
 			})
 
 			It("creates the provided profile", func() {
-				client := fargate.NewClient(clusterName, mockForCreateFargateProfile())
+				client := fargate.NewWithRetryPolicy(clusterName, mockForCreateFargateProfile(), &retryPolicy)
 				waitForCreation := false
 				err := client.CreateProfile(testFargateProfile(), waitForCreation)
 				Expect(err).To(Not(HaveOccurred()))
 			})
 
 			It("fails by wrapping the root error with some additional context for clarity", func() {
-				client := fargate.NewClient(clusterName, mockForFailureOnCreateFargateProfile())
+				client := fargate.NewWithRetryPolicy(clusterName, mockForFailureOnCreateFargateProfile(), &retryPolicy)
 				waitForCreation := false
 				err := client.CreateProfile(testFargateProfile(), waitForCreation)
 				Expect(err).To(HaveOccurred())
@@ -56,7 +58,7 @@ var _ = Describe("fargate", func() {
 					Time: 0, TimeUnit: time.Second, MaxRetries: 5,
 				}
 				numRetriesAfterCreation := 3 // < MaxRetries
-				client := fargate.NewClientWithRetryPolicy(clusterName, mockForCreateFargateProfileWithWait(numRetriesAfterCreation), retryPolicy)
+				client := fargate.NewWithRetryPolicy(clusterName, mockForCreateFargateProfileWithWait(numRetriesAfterCreation), retryPolicy)
 				waitForCreation := true
 				err := client.CreateProfile(testFargateProfile(), waitForCreation)
 				Expect(err).To(Not(HaveOccurred()))
@@ -68,7 +70,7 @@ var _ = Describe("fargate", func() {
 					Time: 0, TimeUnit: time.Second, MaxRetries: 5,
 				}
 				numRetriesAfterCreation := 5 // == MaxRetries, i.e. we will time out.
-				client := fargate.NewClientWithRetryPolicy(clusterName, mockForCreateFargateProfileWithWait(numRetriesAfterCreation), retryPolicy)
+				client := fargate.NewWithRetryPolicy(clusterName, mockForCreateFargateProfileWithWait(numRetriesAfterCreation), retryPolicy)
 				waitForCreation := true
 				err := client.CreateProfile(testFargateProfile(), waitForCreation)
 				Expect(err).To(HaveOccurred())
@@ -78,7 +80,7 @@ var _ = Describe("fargate", func() {
 
 		Describe("ReadProfiles", func() {
 			It("returns all Fargate profiles", func() {
-				client := fargate.NewClient(clusterName, mockForReadProfiles())
+				client := fargate.NewWithRetryPolicy(clusterName, mockForReadProfiles(), &retryPolicy)
 				out, err := client.ReadProfiles()
 				Expect(err).To(Not(HaveOccurred()))
 				Expect(out).To(Not(BeNil()))
@@ -88,7 +90,7 @@ var _ = Describe("fargate", func() {
 			})
 
 			It("returns an empty array if no Fargate profile exists", func() {
-				client := fargate.NewClient(clusterName, mockForEmptyReadProfiles())
+				client := fargate.NewWithRetryPolicy(clusterName, mockForEmptyReadProfiles(), &retryPolicy)
 				out, err := client.ReadProfiles()
 				Expect(err).To(Not(HaveOccurred()))
 				Expect(out).To(Not(BeNil()))
@@ -96,7 +98,7 @@ var _ = Describe("fargate", func() {
 			})
 
 			It("fails by wrapping the root error with some additional context for clarity", func() {
-				client := fargate.NewClient(clusterName, mockForFailureOnReadProfiles())
+				client := fargate.NewWithRetryPolicy(clusterName, mockForFailureOnReadProfiles(), &retryPolicy)
 				out, err := client.ReadProfiles()
 				Expect(err).To(HaveOccurred())
 				Expect(err.Error()).To(Equal("failed to get Fargate profile(s) for cluster \"non-existing-test-cluster\": the Internet broke down"))
@@ -106,14 +108,14 @@ var _ = Describe("fargate", func() {
 
 		Describe("ReadProfile", func() {
 			It("returns the Fargate profile matching the provided name, if any", func() {
-				client := fargate.NewClient(clusterName, mockForReadProfile())
+				client := fargate.NewWithRetryPolicy(clusterName, mockForReadProfile(), &retryPolicy)
 				out, err := client.ReadProfile(testGreen)
 				Expect(err).To(Not(HaveOccurred()))
 				Expect(out).To(Equal(apiFargateProfile(testGreen)))
 			})
 
 			It("returns a 'not found' error if no Fargate profile matched the provided name", func() {
-				client := fargate.NewClient(clusterName, mockForEmptyReadProfile())
+				client := fargate.NewWithRetryPolicy(clusterName, mockForEmptyReadProfile(), &retryPolicy)
 				out, err := client.ReadProfile(testRed)
 				Expect(err).To(HaveOccurred())
 				Expect(err.Error()).To(Equal("failed to get Fargate profile \"test-red\": ResourceNotFoundException: No Fargate Profile found with name: test-red"))
@@ -123,7 +125,7 @@ var _ = Describe("fargate", func() {
 
 		Describe("DeleteProfile", func() {
 			It("fails fast if the provided profile name is empty", func() {
-				client := fargate.NewClient(clusterName, &mocks.EKSAPI{})
+				client := fargate.NewWithRetryPolicy(clusterName, &mocks.EKSAPI{}, &retryPolicy)
 				waitForDeletion := false
 				err := client.DeleteProfile("", waitForDeletion)
 				Expect(err).To(HaveOccurred())
@@ -132,7 +134,7 @@ var _ = Describe("fargate", func() {
 
 			It("deletes the profile corresponding to the provided name", func() {
 				profileName := "test-green"
-				client := fargate.NewClient(clusterName, mockForDeleteFargateProfile(profileName))
+				client := fargate.NewWithRetryPolicy(clusterName, mockForDeleteFargateProfile(profileName), &retryPolicy)
 				waitForDeletion := false
 				err := client.DeleteProfile(profileName, waitForDeletion)
 				Expect(err).To(Not(HaveOccurred()))
@@ -140,7 +142,7 @@ var _ = Describe("fargate", func() {
 
 			It("fails by wrapping the root error with some additional context for clarity", func() {
 				profileName := "test-green"
-				client := fargate.NewClient(clusterName, mockForFailureOnDeleteFargateProfile(profileName))
+				client := fargate.NewWithRetryPolicy(clusterName, mockForFailureOnDeleteFargateProfile(profileName), &retryPolicy)
 				waitForDeletion := false
 				err := client.DeleteProfile(profileName, waitForDeletion)
 				Expect(err).To(HaveOccurred())
@@ -154,7 +156,7 @@ var _ = Describe("fargate", func() {
 					Time: 0, TimeUnit: time.Second, MaxRetries: 5,
 				}
 				numRetriesBeforeDeletion := 3 // < MaxRetries
-				client := fargate.NewClientWithRetryPolicy(clusterName, mockForDeleteFargateProfileWithWait(profileName, numRetriesBeforeDeletion), retryPolicy)
+				client := fargate.NewWithRetryPolicy(clusterName, mockForDeleteFargateProfileWithWait(profileName, numRetriesBeforeDeletion), retryPolicy)
 				waitForDeletion := true
 				err := client.DeleteProfile(profileName, waitForDeletion)
 				Expect(err).To(Not(HaveOccurred()))
@@ -167,7 +169,7 @@ var _ = Describe("fargate", func() {
 					Time: 0, TimeUnit: time.Second, MaxRetries: 5,
 				}
 				numRetriesBeforeDeletion := 5 // == MaxRetries, i.e. we will time out.
-				client := fargate.NewClientWithRetryPolicy(clusterName, mockForDeleteFargateProfileWithWait(profileName, numRetriesBeforeDeletion), retryPolicy)
+				client := fargate.NewWithRetryPolicy(clusterName, mockForDeleteFargateProfileWithWait(profileName, numRetriesBeforeDeletion), retryPolicy)
 				waitForDeletion := true
 				err := client.DeleteProfile(profileName, waitForDeletion)
 				Expect(err).To(HaveOccurred())
